fix conn + up bdd
