+++ client/src/interface/index.js
@@ -3,3 +3,5 @@ export * from "./cut_address"
 export * from "./need_s"
 export * from "./create_list_person"
 export * from "./input_handler_search"
+export * from "./hooks_protected_routes"
+export * from "./update_context"
